LayoutTests/http/tests/media/media-source/video-media-source-play-expected.txt

Tests normal playback case with MediaSource API

EVENT(webkitsourceopen)
EVENT(webkitsourceended)
EVENT(playing)
EVENT(ended)
END OF TEST
